/!\ NON COMPILABLE STATE /!\ Refactoring++ /!\ NON COMPILABLE STATE /!\

+++ YAPM/Providers/Classes/Handles/asyncMethods/asyncCallbackHandleUnload.vb
@@ -29,7 +29,7 @@ Public Class asyncCallbackHandleUnload
     Private con As cHandleConnection
     Private _deg As HasUnloadedHandle
 
-    Public Delegate Sub HasUnloadedHandle(ByVal Success As Boolean, ByVal pid As Integer, ByVal handle As Integer, ByVal msg As String, ByVal actionNumber As Integer)
+    Public Delegate Sub HasUnloadedHandle(ByVal Success As Boolean, ByVal pid As Integer, ByVal handle As IntPtr, ByVal msg As String, ByVal actionNumber As Integer)
 
     Public Sub New(ByVal deg As HasUnloadedHandle, ByRef procConnection As cHandleConnection)
         _deg = deg
